pyproject.toml
[tool.poetry]
name = "faith-echo"
version = "0.1.0"
description = "Real-time Swedish to English/French translation service"
authors = ["FaithEcho Contributors <noreply@example.com>"]
readme = "README.md"
packages = [{ include = "faith_echo", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.73.1"
fastapi = "^0.116.1"
uvicorn = "==0.29.0"
protobuf = "^6.31.1"
google-cloud-speech = "^2.26.0"
google-cloud-translate = "^3.21.1"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.4"
pytest = "^8.4.1"
mypy = "^1.17.0"
pydantic = "^2.11.7"
pip-audit = "^2.9.0"
pre-commit = "^3.7.0"
grpcio-tools = "^1.73.1"
shellingham = "^1.5.4"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.mypy]
python_version = "3.12"
strict = true
